# docs/swagger.yml
openapi: 3.0.0
info:
  title: 'Frontastic Catwalk'
  version: 1.0.0
paths:
...
  /api/cart/cart/checkout:
    post:
      summary: 'Convert a (complete) cart into an order'
      description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: 'null'
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  _type:
                    type: string
                    pattern: ^object$
                  order:
                    type: \Frontastic\Catwalk\FrontendBundle\Controller\Order
...